GeometryCompressionUtils: convert to esmodules
